fixture 파라미터 타입 수정

+++ src/test/java/com/seeyoungryu/connecti/fixture/TestInfoFixture.java
@@ -7,7 +7,7 @@ public class TestInfoFixture {
 
     public static TestInfo get() {
         TestInfo info = new TestInfo();
-        info.setPostId(1);
+        info.setPostId(1L);
         info.setUserName("name");
         info.setPassword("password");
         info.setTitle("title");
@@ -17,7 +17,7 @@ public class TestInfoFixture {
 
     @Data
     public static class TestInfo {
-        private Integer postId;
+        private Long postId;
         private String userName;
         private String password;
         private String title;
